+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # [v0.0.31](https://github.com/opensafely-actions/cox-ipw/releases/tag/v0.0.31)
 
-- The `readr::write_csv()` calls are now more robust to the version of the readr package installed in the r image, i.e., if the version of readr is less that 1.4.0 then the `path` argument is used, otherwise the `file` argument is used. [Tidyverse blogpost about readr 1.4.0](https://www.tidyverse.org/blog/2020/10/readr-1-4-0/#argument-name-consistency).
+- Save analysis ready now uses `foreign::write.dta` so that output can be read directly into Stata
 
 # [v0.0.30](https://github.com/opensafely-actions/cox-ipw/releases/tag/v0.0.30)